requirements.txt
fastapi>=0.95
uvicorn[standard]>=0.20.0
pypdf>=4.0.0
python-docx>=0.8.11
aiofiles>=23.0.0
python-multipart==0.0.6
pillow
pymupdf
langgraph==0.0.26
langchain>=0.1.0
langchain-openai>=0.0.2
qdrant-client>=1.7.0
mcp
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator
pydantic-settings==2.1.0
redis==5.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
elasticsearch==8.11.0
